+++ PS5Parser.cpp
@@ -37,8 +37,8 @@ enum DPADEnum {
 //#define PRINTREPORT // Uncomment to print the report send by the PS5 Controller
 
 int8_t PS5Parser::getButtonIndex(ButtonEnum b) {
-    const uint8_t index = legacyButtonValues(b);
-    if (index >= sizeof(PS5_BUTTONS) / sizeof(PS5_BUTTONS[0])) return -1;
+    const int8_t index = legacyButtonValues(b);
+    if ((uint8_t) index >= (sizeof(PS5_BUTTONS) / sizeof(PS5_BUTTONS[0]))) return -1;
     return index;
 }
